Added a sync parameter to the process method.
A sync parameter has been added to the process method. This will allow to allow a user to turn off the Visual Studio sync. This has been added in to deal with an issue when writing files to a relative directory. The files were added as nested files under the generator and not added to the project. The files were correctly shown as missing underneath generator. This change allows the user to hand manage the files that are added or removed from the project.